Add 'ui.panel' subpackage
