+++ src/edu/colorado/csdms/wmt/client/ui/SignInScreen.java
@@ -19,7 +19,7 @@ import com.google.gwt.user.client.ui.VerticalPanel;
 
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
-import edu.colorado.csdms.wmt.client.ui.widgets.SignInInfoPanel;
+import edu.colorado.csdms.wmt.client.ui.panel.SignInInfoPanel;
 
 /**
  * The {@link SignInScreen} is the first thing a user sees on loading WMT. It
