+++ src/edu/colorado/csdms/wmt/client/ui/panel/SignInInfoPanel.java
@@ -1,4 +1,4 @@
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.panel;
 
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.ClickHandler;
