+++ src/edu/colorado/csdms/wmt/client/ui/panel/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Panels are composites of smaller UI elements used in the WMT client
+ * interface.
+ * 
+ * @author Mark Piper (mark.piper@colorado.edu) 
+ */
+package edu.colorado.csdms.wmt.client.ui.panel;
