+++ src/edu/colorado/csdms/wmt/client/ui/widgets/NewUserDialogBox.java
@@ -30,6 +30,8 @@ import com.google.gwt.user.client.ui.HasHorizontalAlignment;
 import com.google.gwt.user.client.ui.VerticalPanel;
 
 import edu.colorado.csdms.wmt.client.Constants;
+import edu.colorado.csdms.wmt.client.ui.panel.ChoicePanel;
+import edu.colorado.csdms.wmt.client.ui.panel.FieldPanel;
 
 /**
  * A {@link DialogBox} that prompts a user to reenter their password when
