+++ src/edu/colorado/csdms/wmt/client/ui/widgets/SaveDialogBox.java
@@ -34,6 +34,8 @@ import com.google.gwt.user.client.ui.VerticalPanel;
 
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
+import edu.colorado.csdms.wmt.client.ui.panel.ChoicePanel;
+import edu.colorado.csdms.wmt.client.ui.panel.FieldPanel;
 
 /**
  * A customized DialogBox with a field for setting the name/description of the
